Teleporters work again, audio works, audio settings work, organized folders
Teleporters needed the game instance to be set as it was set to none so it wasn't working as intended, added audio to the game and the settings change the sounds in real time and save the settings that you put. Made a particle folder and a sounds folder and both well organized with everything we need.